test: reuse dedup fixtures for exact sequence composition
