+++ src/mito2/src/read/series_reader.rs
@@ -598,137 +598,6 @@ mod tests {
     }
 
     #[test]
-    fn exact_sequence_filter_component_composition_covers_tombstones_and_last_non_null() {
-        // This component-composition test applies the row-level sequence filter
-        // before the shared dedup iterator; it does not construct production
-        // series-reader wiring. It complements integration coverage with
-        // tombstone and LastNonNull behavior.
-        use api::v1::OpType;
-        use datatypes::arrow::array::{
-            ArrayRef, BinaryDictionaryBuilder, Int64Array, StringArray, TimestampMillisecondArray,
-            UInt8Array, UInt64Array,
-        };
-        use datatypes::arrow::datatypes::{DataType, Field, Schema, UInt32Type};
-        use datatypes::arrow::record_batch::RecordBatch;
-        use store_api::storage::SequenceRange;
-
-        use crate::read::flat_dedup::{FlatDedupIterator, FlatLastNonNull};
-
-        let schema = Arc::new(Schema::new(vec![
-            Field::new("tag_0", DataType::Utf8, false),
-            Field::new("field_0", DataType::Int64, true),
-            Field::new("field_1", DataType::Int64, true),
-            Field::new(
-                "ts",
-                DataType::Timestamp(datatypes::arrow::datatypes::TimeUnit::Millisecond, None),
-                false,
-            ),
-            Field::new(
-                "__primary_key",
-                DataType::Dictionary(Box::new(DataType::UInt32), Box::new(DataType::Binary)),
-                false,
-            ),
-            Field::new("__sequence", DataType::UInt64, false),
-            Field::new("__op_type", DataType::UInt8, false),
-        ]));
-        let batch = |timestamp: i64,
-                     fields: &[(Option<i64>, Option<i64>)],
-                     sequences: &[u64],
-                     ops: &[OpType]| {
-            let mut keys = BinaryDictionaryBuilder::<UInt32Type>::new();
-            for _ in fields {
-                keys.append(b"series").unwrap();
-            }
-            RecordBatch::try_new(
-                schema.clone(),
-                vec![
-                    Arc::new(StringArray::from_iter_values(
-                        (0..fields.len()).map(|_| "series"),
-                    )) as ArrayRef,
-                    Arc::new(Int64Array::from_iter(fields.iter().map(|field| field.0))) as ArrayRef,
-                    Arc::new(Int64Array::from_iter(fields.iter().map(|field| field.1))) as ArrayRef,
-                    Arc::new(TimestampMillisecondArray::from_iter_values(
-                        (0..fields.len()).map(|_| timestamp),
-                    )) as ArrayRef,
-                    Arc::new(keys.finish()) as ArrayRef,
-                    Arc::new(UInt64Array::from_iter_values(sequences.iter().copied())) as ArrayRef,
-                    Arc::new(UInt8Array::from_iter_values(ops.iter().map(|op| *op as u8)))
-                        as ArrayRef,
-                ],
-            )
-            .unwrap()
-        };
-        let filter = |batch| {
-            filter_flat_batch_by_sequence(
-                batch,
-                Some(SequenceRange::GtLtEq { min: 1, max: 3 }),
-                true,
-            )
-            .unwrap()
-            .unwrap()
-        };
-
-        // The in-range tombstone reaches shared dedup and suppresses its put.
-        let output = FlatDedupIterator::new(
-            // Shared merge presents duplicate rows newest-first: the in-range
-            // tombstone (seq 3) must therefore suppress the eligible put
-            // (seq 2).
-            vec![
-                Ok(filter(batch(
-                    1000,
-                    &[(None, None)],
-                    &[3],
-                    &[OpType::Delete],
-                ))),
-                Ok(filter(batch(
-                    1000,
-                    &[(Some(10), None)],
-                    &[2],
-                    &[OpType::Put],
-                ))),
-            ]
-            .into_iter(),
-            FlatLastNonNull::new(1, true),
-        )
-        .collect::<Result<Vec<_>>>()
-        .unwrap();
-        assert!(output.is_empty());
-
-        // A tombstone outside H is removed before shared dedup and cannot
-        // suppress the eligible put.
-        let output = FlatDedupIterator::new(
-            vec![Ok(filter(batch(
-                1000,
-                &[(Some(10), None), (None, None)],
-                &[2, 4],
-                &[OpType::Put, OpType::Delete],
-            )))]
-            .into_iter(),
-            FlatLastNonNull::new(1, true),
-        )
-        .collect::<Result<Vec<_>>>()
-        .unwrap();
-        assert_eq!(1, output[0].num_rows());
-
-        // An outside-range non-null value is removed before LastNonNull, so it
-        // cannot fill the eligible row's null field.
-        let output = FlatDedupIterator::new(
-            vec![Ok(filter(batch(
-                1000,
-                &[(Some(10), None), (None, Some(40))],
-                &[2, 4],
-                &[OpType::Put, OpType::Put],
-            )))]
-            .into_iter(),
-            FlatLastNonNull::new(1, true),
-        )
-        .collect::<Result<Vec<_>>>()
-        .unwrap();
-        assert_eq!(1, output[0].num_rows());
-        assert!(output[0].column(2).is_null(0));
-    }
-
-    #[test]
     fn series_range_assignment_is_stable_across_batch_boundaries() {
         let input = vec![
             series(1, 0),
